// src/api/generated/models/householdData.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Rukun API
 * Standard API contract for Core R.
 * OpenAPI spec version: 1.0.0
 */
import type { HouseholdDataBlock } from './householdDataBlock';
import type { HouseholdDataHouseNumber } from './householdDataHouseNumber';
import type { HouseholdDataOccupancyStatus } from './householdDataOccupancyStatus';

export interface HouseholdData {
  public_id?: string;
  reference?: string;
  area_id?: string;
  status?: string;
  /** @maxLength 2000 */
  address?: string;
  /** @maxLength 50 */
  block?: HouseholdDataBlock;
  /** @maxLength 50 */
  house_number?: HouseholdDataHouseNumber;
  occupancy_status?: HouseholdDataOccupancyStatus;
}