// src/api/generated/models/moveResident.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Rukun API
 * Standard API contract for Core R.
 * OpenAPI spec version: 1.0.0
 */
import type { MoveResidentHouseholdId } from './moveResidentHouseholdId';
import type { MoveResidentRelationship } from './moveResidentRelationship';

export interface MoveResident {
  household_id: MoveResidentHouseholdId;
  relationship: MoveResidentRelationship;
}